requirements.txt
dash==2.16.1
plotly==5.18.0
numpy==1.24.3
dash-bootstrap-components==1.5.0
pandas==2.1.4
scipy==1.11.4
werkzeug==3.0.6
pyyaml==6.0.1
gunicorn==22.0.0

# Optional: UMAP computation (only needed to recompute position embeddings)
# umap-learn